kql-expert-chatbot/pyproject.toml
[tool.poetry]
name = "kql-expert-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kql-expert-chatbot"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
langchain = "^0.0.161"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
pbyc = {path = "../../Library"}
aiohttp = "^3.8.4"
PyMuPDF = "^1.22.1"
docx2txt = "^0.8"
tiktoken = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
